rename notifications tab to charts
