+++ app/src/main/java/com/example/oddhours/ui/charts/ChartsViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.oddhours.ui.notifications
+package com.example.oddhours.ui.charts
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class ChartsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is the charts Fragment"
     }
     val text: LiveData<String> = _text
 }
