+++ app/src/main/res/layout/fragment_charts.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.notifications.NotificationsFragment">
+    tools:context=".ui.charts.ChartsFragment">
 
     <TextView
-        android:id="@+id/notificationTv"
+        android:id="@+id/chartTv"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
